*)Initial values of all parameters for all slaves are taken from the database in the beginning.(If they are not available in db,they are set to 0). Then for further database updation,the current value is compared with this value and if they differ,new value is added in db.
*)Initially when there is no value for any parameter for any slave device,
all values are set to 0,and data is updated as it is received from the meter.

*)After the initial setting,the database updation takes place only if the
present value differs from the value in the db by a preset offset
amount(maybe 1% or 5%,have to decide this..).This offset is set as follows.
If the live value is not zero,the offset is set to be 1% of the
live value(otherwise it is set as 1).Then for database updation the
previous value and live value are compared to see if they differ by this
offset.If they differ,the new value is added to DB.