Following changes are made in this commit to reflect the changes done in Kims Code.Later when checking out,this should be compared to Kims code to see if there is any difference.
1)In ProcessParamsA,B,C functions,the 2 lines relating to updation of params
in db based on params specified in config file was commented.
So now all params from meter will be updated in DB.
2)For cumulative energy field and related fields,the setting of offset to 1%
of prev value was commented out in ProcessParams B and C functions
3)database name was changed to ilids_nov instead of ilids.

Also username and email was configured newly using command git config,as this
commit is done from home laptop.