Major Update : Selective updation of DB based on only the params specified in config file is enabled. Initially all the params are set to be updated.But if VALID_PARAMS tag is present in config file,only those params specified under it will be updated to DB.
Valid Params are specified as follows in the config file :

[VALID_PARAMS]
7:(0,1),(6,2),(30,0.05)
8:(2,2),(8,3),(34,0.10)
9:(4,2),(10,4),(32,0.05)

Here the first field is the slaveid.Then after the column,the list of
tags to be updated are specified inside parantheses.Within the parantheses,
the address and offset of the params are specified in that order.
Effectively it will come as
[VALID_PARAMS]
slaveid:(addr1,offset1),(addr2,offset2),(addr3,offset3)

For each slaveid,all valid params can be specified like this.
After this,only the Params specified here will be updated in the DB for
that particular slave device.Also the updation happens only if the previous
value differs from the current value by the new specified offset.

Modifications done:
The PARAM_DETAILS structure is added one more field called db_update which
is initially set to 1.Then in readconf function,if valid_params tag is
present,all the db_update flag is reset to 0.After this, the GetValidParams
function sets the db_update flag to 1 for the specified slaves and addresses
only.Then in ProcessParam Functions,the dbupdate flag is checked before
processing the incoming data and the values are updated only if the flag
is set.

Following Files are modified in this update:
id_common.h id_conf.c id_main.c id_proto.c idriver idriver.conf idriver.layout